remoção de modelo de detcao duplicado, mudanca dos ficheiros python do diretorio scripts para a raiz do projeto
