+++ main.py
@@ -1,6 +1,6 @@
 # --------------- BEM VINDO PROFESSOR -----------------------#
 
-from scripts.mediapipe_gesture_recognition import start_server
+from mediapipe_gesture_recognition import start_server
 
 if __name__ == "__main__":
     start_server()
